+++ source/packages/@aws-accelerator/accelerator/test/network-vpc-v2-stacks.test.ts
@@ -36,6 +36,10 @@ const v2TestNamePatterns: { testName: string; stackKey: string }[] = [
     stackKey: 'RouteTableStack-Network-us-east-1-Network-Secondary',
   },
   {
+    testName: 'Construct(RouteTableStack-Network-us-east-1-Network-Inspection-V2): ',
+    stackKey: 'RouteTableStack-Network-us-east-1-Network-Inspection-V2',
+  },
+  {
     testName: 'Construct(SecurityGroupStack-Network-us-east-1-Network-Secondary-V2): ',
     stackKey: 'SecurityGroupStack-Network-us-east-1-Network-Secondary-V2',
   },